# src/Planar/Content/planar_openapi.json
{
  "openapi": "3.0.1",
...
  "info": {
    "title": "Planar",
    "version": "v1",
    "contact": {
      "name": "Tsahi Atias",
      "email": "admin@planar.me",
      "url": "www.planar.me"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "description": "REST API services of Planar system to get information, edit & control schedule task"
  },
  "paths": {
...
    "/group": {
      "post": {
        "tags": [
          "Group"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupRecord"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "AddGroup"
      },
      "get": {
        "tags": [
          "Group"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupInfo"
                  }
                }
              }
            }
          }
        },
        "summary": "GetAllGroups"
      }
    },
    "/group/{id}": {
      "get": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetails"
                }
              }
            }
          }
        },
        "summary": "GetGroup"
      },
      "delete": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        },
        "summary": "DeleteGroup"
      },
      "patch": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityRecord"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          }
        },
        "summary": "UpdateGroup"
      },
      "parameters": [
        {
          "schema": {
            "type": "integer"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ]
    },
    "/group/{id}/user": {
      "post": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserToGroupRecord"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          }
        },
        "summary": "AddUserToGroup"
      },
      "parameters": [
        {
          "schema": {
            "type": "integer"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ]
    },
    "/group/{id}/user/{userId}": {
      "delete": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        },
        "summary": "RemoveUserFromGroup"
      },
      "parameters": [
        {
          "schema": {
            "type": "integer"
          },
          "name": "userId",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
      "AddGroupRecord": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateEntityRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "propertyValue": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserToGroupRecord": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GroupInfo": {
        "title": "GroupInfo",
        "x-stoplight": {
          "id": "yh4a3qeu8uuid"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "userCount": {
            "type": "integer"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "GroupDetails": {
        "title": "GroupDetails",
        "x-stoplight": {
          "id": "ow3lcnoki2b6z"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "reference1": {
            "type": "string"
          },
          "reference2": {
            "type": "string"
          },
          "reference3": {
            "type": "string"
          },
          "reference4": {
            "type": "string"
          },
          "reference5": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:2306",
      "description": "Local"
    }
  ]
}